wip: clean up generalization a bit
